Se limpia el index,creando un nuevo componente. Se hacen las recomendaciones que eslint me dice
